Fix assignee edge cases
I was pulling the first assignee listed but sometimes a security interest assignment is included and should be skipped. The code now looks for the first (which is chronologically most recent) assignment of assignor's interest and then looks for the sibling assignee. I'm also use the str.title() function on the all-caps assignee so I added a replace function to correct Llc to LLC.